chore: re-organize local run
