+++ packages/dashboard-backend/src/localRun/proxies/cheServerApi.ts
@@ -12,6 +12,7 @@
 
 import { FastifyInstance } from 'fastify';
 import fastifyHttpProxy from '@fastify/http-proxy';
+import { stubCheServerOptionsRequests } from '../hooks/stubCheServerOptionsRequests';
 
 export function registerCheApiProxy(
   server: FastifyInstance,
@@ -36,13 +37,5 @@ export function registerCheApiProxy(
       },
     },
   });
-  // stub OPTIONS requests to '/api/' since they fail when running the backend locally.
-  server.addHook('onRequest', (request, reply, done) => {
-    if ((request.url === '/api' || request.url === '/api/') && request.method === 'OPTIONS') {
-      return reply.send({
-        implementationVersion: 'Local Run',
-      });
-    }
-    done();
-  });
+  stubCheServerOptionsRequests(server);
 }
