+++ packages/dashboard-backend/src/localRun/proxies/dexProxies.ts
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2018-2021 Red Hat, Inc.
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Red Hat, Inc. - initial API and implementation
+ */
+
+import { FastifyInstance } from 'fastify';
+import fastifyHttpProxy from '@fastify/http-proxy';
+
+export function registerDexProxies(server: FastifyInstance, dexIngress: string) {
+  server.register(fastifyHttpProxy, {
+    upstream: dexIngress,
+    prefix: '/approval',
+    rewritePrefix: '/approval',
+    disableCache: true,
+  });
+
+  server.register(fastifyHttpProxy, {
+    upstream: dexIngress,
+    prefix: '/theme',
+    rewritePrefix: '/theme',
+    disableCache: true,
+  });
+
+  server.register(fastifyHttpProxy, {
+    upstream: dexIngress,
+    prefix: '/static',
+    rewritePrefix: '/static',
+    disableCache: true,
+  });
+
+  server.register(fastifyHttpProxy, {
+    upstream: dexIngress,
+    prefix: '/token',
+    rewritePrefix: '/token',
+    disableCache: true,
+  });
+
+  server.register(fastifyHttpProxy, {
+    upstream: dexIngress,
+    prefix: '/auth',
+    rewritePrefix: '/auth',
+    disableCache: true,
+  });
+}
